Added thread detach. code cleanup
