--- xarm_api/include/xarm/linux/thread.h
@@ -1,9 +1,0 @@
-/* Copyright 2017 UFACTORY Inc. All Rights Reserved.
- *
- * Software License Agreement (BSD License)
- *
- * Author: Jimy Zhang <jimy92@163.com>
- ============================================================================*/
-#ifndef LINUX_THREAD_H_
-#define LINUX_THREAD_H_
-#endif
